requirements.txt
# Python dependencies for execution scripts
python-dotenv>=1.0.0

# Google APIs (uncomment when needed)
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-api-python-client>=2.0.0

# Common utilities (add as needed)
requests>=2.28.0
pandas>=2.0.0
supabase>=2.0.0
gspread>=5.10.0
google-auth>=2.0.0
lseg-data>=1.0.0
